Dev/robin/8336 signed merkle roots 4 (#49)
* MassifContext can be used with GetRoot

This PR adds the CreatePeaksStackMap method to the MassifContext.

The default Add method on the MassifContext is optimised for the Add
leaf case. As this is the hot path. GetRoot accesses the mmr
differently, so when the context is used for GetRoot it needs a more
explicit picture of the peaks carried over from previous massifs

The CreatePeaksStackMap method should be called before using the context
with GetRoot. The MassifReader type has gained With option support to
it's New method so that it can be configured easily for this case.

AB#8336

* Signed confirmation of log state

* implements a logconfirmer which periodically verifies the consistency
  of the log and signs the latest root.
* Every massif backing the log is associated with a single 'Signed Tree
  Head'. While the massif is being filled, its just the most recent
  head. Once the massif is complete, the logconfirmer will sign the
  root corresponding to the MMRSize at the end of that massif. There
  after, that published signed tree head will never change.
* Also, rename triecommon -> merklelog

AB# 8336

* linter issues

* Review feedback & fix issue with mapped peak stack

* remove duplicated logging

---------

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>