blind the trie index data (#75)
This change replaces the trie index data fields, which included the
event identities, with opaque hashes.

H(DOMAIN, idtimestamp, applicationIndex)

applicationIndex is just our eventIdentity *string* as bytes

We previously chose to reveal the event identities in the trie index
data. We understood this leaked the cross tenant activity. We believed
it was hard to do the CONFIRMED status without having cached copies of
the CONFIRMED event identities in the forestrie blobs.

It turns out that,

a) Generaly allowing side channel analysis of a single tenants frequency
of publishing is still OK
b) Generaly allowing side channel analysis of sharing accross tenants is
NOT OK (this pr fixes this issue)
c) The way confirmation came out, keeping the event identities in the
log data isn't necessary at all.

This does mean that demos showing non equivocation and completeness will
need to have pre-image data for events if they are to match suspect
entries up against real api data. The merklelogentry.commit.index should
be used to derive the leaf slot, then  H(domain, idtimestamp,
eventIdentity)  should be calculated and compared against what is
actually in the log.

Similarly, we need to be able to do MMR-INDEX-OF(LEAF-INDEX) -> event
api query on ?merklelogentry.commit.index in order to go from log to api
data. That will be added post v1 release.

In general, auditors of the logs would need to be given pre-image
data of *just* the idtimestamp and the event identity.

The applicaiton is responsible for including the applicationIndex in
their leaf hash, to ensure the trie index is commited by the log. And
our simplehash schema does so.

Additionaly this PR,

* Delete some (now) redundant code.
* fixes a bug in how we save & retrieve the trie index value, we were
  using mmrIndex where we should have been using leafIndex
* reserves 7 32 byte fields for future extension. the first use is
  anticipated to be for the urkle trie root
* Adds integration tests to cover the bug fix
* REMOVES the byte type code from the start header. we will rely
  exclusively on the VERSION field in the header.

AB#9152

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>