fix: 10433 remove the path checking (#42)
* fix: 10433 remove the path checking

The path checking was intended as an early guard against the case where
a mix of tenant massifs in the same directory would cause replication to
fail. The intent was to help the caller by providing them and indication
of the problematic file names

Instead, we now have tests (in veracity) which show the explicit errors
for this case are surfaced from this library and are acted on
appropriately.

* fix: test case that expected the duplicate massif based on filename error

* fix: apply same fix for the seal cache

---------

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>